% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fairy_tales}
\alias{fairy_tales}
\title{Sample of 5 fairly tales from H.C. Andersen}
\format{
A tibble with 486 obs. of 2 variables.
}
\usage{
fairy_tales
}
\description{
Includes the fairy tales "A leaf from heaven", "A story",
"The bird of folklore", "The flea and the professor" and "The rags".
}
\keyword{datasets}
