% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_coef_plot.R
\name{glmnet_coef_plot}
\alias{glmnet_coef_plot}
\title{Create Lambda chart for glmnet object}
\usage{
glmnet_coef_plot(fit)
}
\arguments{
\item{fit}{parsnip fit object}
}
\value{
ggplot2 object
}
\description{
Create Lambda chart for glmnet object
}
\examples{
library(parsnip)
library(glmnet)
linear_reg_glmnet_spec <-
  linear_reg(penalty = 0, mixture = 1) \%>\%
  set_engine('glmnet')

lm_fit <- fit(linear_reg_glmnet_spec, data = mtcars, mpg ~ .)

glmnet_coef_plot(lm_fit)
}
