% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnist_sample}
\alias{mnist_sample}
\title{Down-sampled MNIST}
\format{
A data frame with 1200 rows and 3 variables:
\describe{
  \item{class}{Factor, values 0 through 9.}
  \item{umap_1}{Numeric.}
  \item{umap_2}{Numeric.}
}
}
\source{
\url{http://yann.lecun.com/exdb/mnist/}
}
\usage{
mnist_sample
}
\description{
this data set is a down-sampled version of the MNIST data set which have been
umap'ed down to 2 dimensions.
}
\keyword{datasets}
