% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_classbalance.R
\name{viz_classbalance}
\alias{viz_classbalance}
\title{Viszualise}
\usage{
viz_classbalance(data, variable, n_max = 25)
}
\arguments{
\item{data}{A data.frame.}

\item{variable}{target variable to show balance for.}

\item{n_max}{integer, maximum number of classes shown before lumping.
Defaults to 25.}
}
\value{
ggplot2 object.
}
\description{
Viszualise
}
\examples{
viz_classbalance(mtcars, vs)
}
