% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_decision_boundary.R
\name{viz_decision_boundary}
\alias{viz_decision_boundary}
\title{Title}
\usage{
viz_decision_boundary(x, new_data, grid = TRUE, resolution = 100, expand = 0.1)
}
\arguments{
\item{x}{fitted workflow object}

\item{new_data}{data frame}

\item{grid}{Logical, whether to use grid method to show decision boundary.
Defaults to TRUE.}

\item{resolution}{Number of squared in grid. Defaults to 100.}

\item{expand}{Expansion rate. Defaults to 0.1. This means that the width and
height of the shaded area is 10% wider then the rectangle containing the
data.}
}
\value{
ggplot object
}
\description{
Title
}
\examples{
library(parsnip)
library(workflows)
lm_spec <- svm_rbf() \%>\%
  set_mode("classification") \%>\%
  set_engine("kernlab")

lr_fit <- workflow() \%>\%
  add_formula(Species ~ Petal.Length + Petal.Width) \%>\%
  add_model(lm_spec) \%>\%
  fit(tidyr::drop_na(iris))

viz_decision_boundary(lr_fit, iris)
}
