% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_pca.R
\name{viz_pca}
\alias{viz_pca}
\title{Vizualize principal pomponents}
\usage{
viz_pca(data, label, components = c(1, 2), loadings = FALSE)
}
\arguments{
\item{data}{A data.frame.}

\item{label}{variable to color with.}

\item{components}{principal components to showcase.}

\item{loadings}{Set this to true if you want to see the PCA loadings.}
}
\value{
ggplot2 object.
}
\description{
Vizualize principal pomponents
}
\examples{
viz_pca(iris, Species)
viz_pca(iris, Species, c(3, 1))
viz_pca(iris, Species, loadings = TRUE)
}
