% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_pcacm.R
\name{viz_pcacm}
\alias{viz_pcacm}
\title{PCA component measures}
\usage{
viz_pcacm(x, n_pca = 10, n_var = 10)
}
\arguments{
\item{x}{a data.frame}

\item{n_pca}{Number of Principle components to show}

\item{n_var}{Number of Variables to show}
}
\value{
a ggplot2 object
}
\description{
PCA component measures
}
\examples{
viz_pcacm(mtcars)
viz_pcacm(USArrests)
viz_pcacm(beaver1)
}
