% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_prob_distribution.R
\name{viz_prob_distribution}
\alias{viz_prob_distribution}
\alias{viz_prob_distribution.default}
\alias{viz_prob_distribution.model_fit}
\title{Vizualize the predicted probalities for each class}
\usage{
viz_prob_distribution(model, new_data, truth)

\method{viz_prob_distribution}{default}(model, new_data, truth)

\method{viz_prob_distribution}{model_fit}(model, new_data, truth = NULL)
}
\arguments{
\item{model}{A `model_fit` object from the parsnip package.}

\item{new_data}{A data.frame to run predictions on.}

\item{truth}{A vector of true classes used to color probalities.
Defaults to NULL.}
}
\value{
ggplot2 object.
}
\description{
Vizualize the predicted probalities for each class
}
\examples{
library(parsnip)
library(ranger)

fit_model <- rand_forest("classification") \%>\%
  set_engine("ranger") \%>\%
  fit(Species ~ ., data = iris)

viz_prob_distribution(fit_model, new_data = iris)

viz_prob_distribution(fit_model, new_data = iris, truth = iris$Species)
}
