% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_prob_region.R
\name{viz_prob_region}
\alias{viz_prob_region}
\title{Draw Probability regions for Classification model}
\usage{
viz_prob_region(x, new_data, resolution = 100, expand = 0.1)
}
\arguments{
\item{x}{trained `workflows::workflow` object.}

\item{new_data}{A data frame or tibble for whom the preprocessing will be
applied.}

\item{resolution}{Number of squared in grid. Defaults to 100.}

\item{expand}{Expansion rate. Defaults to 0.1. This means that the width and
  height of the shaded area is 10% wider then the rectangle containing the
  data.

The chart have been minimally modified to allow for easier styling.}
}
\value{
`ggplot2::ggplot` object
}
\description{
This function is mostly useful in an educational setting. Can only be used
with trained workflow objects with 2 numeric predictor variables.
}
\examples{
library(parsnip)
library(workflows)

iris2 <- iris
iris2$Species <- factor(iris2$Species == "setosa",
                        labels = c("setosa", "not setosa"))

svm_spec <- svm_rbf() \%>\%
  set_mode("classification") \%>\%
  set_engine("kernlab")

svm_fit <- workflow() \%>\%
  add_formula(Species ~ Petal.Length + Petal.Width) \%>\%
  add_model(svm_spec) \%>\%
  fit(iris2)

viz_prob_region(svm_fit, iris2)

viz_prob_region(svm_fit, iris2, resolution = 20)

viz_prob_region(svm_fit, iris2, expand = 1)
}
