% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_residuals.R
\name{viz_residuals}
\alias{viz_residuals}
\title{Residual plot of model}
\usage{
viz_residuals(fit, data)
}
\arguments{
\item{fit}{a parsnip model object}

\item{data}{data.frame to create residuals with}
}
\value{
a ggplot2 object
}
\description{
Residual plot of model
}
\examples{
library(parsnip)
reg <- linear_reg() \%>\%
  set_engine("lm") \%>\%
  set_mode("regression") \%>\%
  fit(mpg ~ ., data = mtcars)

viz_residuals(reg, mtcars)
}
