% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tsne.R
\name{viz_tsne}
\alias{viz_tsne}
\title{Vizualize t-SNE}
\usage{
viz_tsne(data, label)
}
\arguments{
\item{data}{A data.frame.}

\item{label}{variable to color with.}
}
\value{
ggplot2 object.
}
\description{
Vizualize t-SNE
}
\examples{
viz_tsne(iris, Species)
viz_tsne(dplyr::select(iris, -Species), Sepal.Length)
}
