% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_decision_boundary.R
\name{viz_decision_boundary}
\alias{viz_decision_boundary}
\title{Draw Decision boundary for Classification model}
\usage{
viz_decision_boundary(x, new_data, resolution = 100, expand = 0.1)
}
\arguments{
\item{x}{trained `workflows::workflow` object.}

\item{new_data}{A data frame or tibble for whom the preprocessing will be
applied.}

\item{resolution}{Number of squared in grid. Defaults to 100.}

\item{expand}{Expansion rate. Defaults to 0.1. This means that the width and
  height of the shaded area is 10% wider then the rectangle containing the
  data.

The chart have been minimally modified to allow for easier styling.}
}
\value{
`ggplot2::ggplot` object
}
\description{
This function is mostly useful in an educational setting. Can only be used
with trained workflow objects with 2 numeric predictor variables.
}
\examples{
library(parsnip)
library(workflows)
svm_spec <- svm_rbf() \%>\%
  set_mode("classification") \%>\%
  set_engine("kernlab")

svm_fit <- workflow() \%>\%
  add_formula(Species ~ Petal.Length + Petal.Width) \%>\%
  add_model(svm_spec) \%>\%
  fit(iris)

viz_decision_boundary(svm_fit, iris)
}
