% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_dispersion.R
\name{viz_dispersion}
\alias{viz_dispersion}
\title{Vizualize lexical dispersion plot}
\usage{
viz_dispersion(data, var, words = 10)
}
\arguments{
\item{data}{A data.frame.}

\item{var}{variable that contains the words}

\item{words}{Numerical or character. If numerical it will display the n
most common words. If character will show the location of said strings.}
}
\value{
ggplot2 object.
}
\description{
Vizualize lexical dispersion plot
}
\examples{

library(tidytext)

text_data <- tidytext::unnest_tokens(fairy_tales, text, text)
viz_dispersion(text_data, text)
viz_dispersion(text_data, text, "kissed")
}
